requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Limiter==3.5.0

# Database
SQLAlchemy==2.0.21
psycopg[binary]==3.1.10
alembic==1.12.0

# Task processing
celery==5.3.1
redis==4.6.0

# Validation and serialization
pydantic==2.3.0
marshmallow==3.20.1

# HTTP client for testing
requests==2.31.0

# Environment management
python-dotenv==1.0.0

# Utilities
click==8.1.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
Werkzeug==2.3.7

# Development dependencies
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
black==23.7.0
isort==5.12.0
flake8==6.0.0

# WSGI server for production
gunicorn==21.2.0

# Process management
honcho==1.1.0